rename all auth-classes and types to session-
