+++ src/app/api/session/session.iservice.ts
@@ -8,7 +8,7 @@ import { Person } from 'src/app/models/entities/Person';
 import { User } from 'src/app/models/entities/User';
 import { AuthorizedAccess } from 'src/app/models/AuthorizedAccess';
 
-export interface AuthenticationIService {
+export interface SessionIService {
 
   guestLogin(details: Person): Observable<boolean>;
   register(details: Partial<User>): Observable<boolean>;
