+++ src/app/management/routes/users/user-manager.access-resolver.ts
@@ -7,7 +7,7 @@ import { Inject, Injectable } from '@angular/core';
 import { ActivatedRouteSnapshot, Resolve, RouterStateSnapshot } from '@angular/router';
 import { Observable } from 'rxjs';
 import { AUTH_INJECTION_TOKEN } from 'src/app/api/session/auth.injection-token';
-import { AuthenticationIService } from 'src/app/api/session/auth.iservice';
+import { SessionIService } from 'src/app/api/session/session.iservice';
 import { AuthorizedAccess } from 'src/app/models/AuthorizedAccess';
 
 @Injectable({ providedIn: 'root' })
@@ -15,7 +15,7 @@ export class UserManagerAccessResolver
   implements Resolve<AuthorizedAccess> {
 
   constructor(
-    @Inject(AUTH_INJECTION_TOKEN) protected authService: AuthenticationIService
+    @Inject(AUTH_INJECTION_TOKEN) protected authService: SessionIService
   ) { }
 
   resolve(
